# Run the remote image presence checks in parallel (speeds up pushes)