@jfrazelle 2015-01-13
Hi @IanCal,
We are going to close this becuase there have not been any changes to the implementation. But we do like the idea. So if you get time to re-implement with maybe a channel, or something less racy. We will gladly reopen. Thanks!