@crosbymichael 2015-01-13
I'll rebase this one can fix some of the concurrency issues