@crosbymichael 2015-01-13
Closing in favor of #10077

@IanCal if you can, can you try running your performance checks on this new PR?