@IanCal 2015-01-14
That sounds like a good idea, sorry I've not had the time recently to work
on it. I'll hopefully have something in a week or two when I'm back in the
country if someone else hasn't done it in the meantime :)

On Tue, 13 Jan 2015 19:47 Jessie Frazelle <notifications@github.com> wrote:

> Hi @IanCal <https://github.com/IanCal>,
> We are going to close this becuase there have not been any changes to the
> implementation. But we do like the idea. So if you get time to re-implement
> with maybe a channel, or something less racy. We will gladly reopen. Thanks!
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/docker/docker/pull/9714#issuecomment-69796288>.
>
> ********************************************************************************
>
> DISCLAIMER: This e-mail is confidential and should not be used by anyone
> who is
> not the original intended recipient. If you have received this e-mail in
> error
> please inform the sender and delete it from your mailbox or any other
> storage
> mechanism. Neither Macmillan Publishers Limited nor any of its agents
> accept
> liability for any statements made which are clearly the sender's own and
> not
> expressly made on behalf of Macmillan Publishers Limited or one of its
> agents.
> Please note that neither Macmillan Publishers Limited nor any of its agents
> accept any responsibility for viruses that may be contained in this e-mail
> or
> its attachments and it is your responsibility to scan the e-mail and
> attachments (if any). No contracts may be concluded on behalf of Macmillan
> Publishers Limited or its agents by means of e-mail communication.
> Macmillan
> Publishers Limited Registered in England and Wales with registered number
> 785998
> Registered Office Brunel Road, Houndmills, Basingstoke RG21 6XS
> ************************************************************
> ********************
>
>