@crosbymichael 2015-01-14
@IanCal no problem,  I made the changes and It's merged.  Will be in docker 1.5 .  I included your commits also so thanks for the contribution!